created custom year month range

+++ src/main/kotlin/no/nav/pensjon/opptjening/omsorgsopptjening/bestem/pensjonsopptjening/domain/Omsorgsyter.kt
@@ -1,5 +1,5 @@
 package no.nav.pensjon.opptjening.omsorgsopptjening.bestem.pensjonsopptjening.domain
 
 class Omsorgsyter(private val omsorgsArbeidPeriode: List<OmsorgsArbeidPeriode>) {
-    fun monthsOfOmsorg(omsorgsAr: Int): Long = omsorgsArbeidPeriode.firstOrNull()?.monthsInYear(omsorgsAr) ?: 0
+    fun monthsOfOmsorg(omsorgsAr: Int): Int = omsorgsArbeidPeriode.firstOrNull()?.monthsInYear(omsorgsAr) ?: 0
 }
